+++ src/graphics.py
@@ -3,44 +3,43 @@ def display_map(screen, config, level, player):
     tile_legend = level.map.legend
     for map_y in range(0, screen.map_display_height):
         for map_x in range(0, screen.map_display_width):
-            x_offset, y_offset = screen.camera.get_tile_offset(player)
+            x_offset, y_offset = screen.camera.get_offset(level, player)
             current_index = level.map.get_index(
                 map_x - x_offset, map_y - y_offset
             )
-            image_filename = tile_legend.get(unicode(tiles[current_index]), '')
-            image = level.map.images[image_filename]
+            image = tile_legend.get(unicode(tiles[current_index]), '')
             screen.draw_tile(image, (map_x, map_y))
 
 
 def display_player(screen, config, level, player):
-    image = config.images.get(config.player_image, None)
-    screen.draw_tile_relative(image, player, (player.x, player.y))
+    screen.draw_tile_relative(
+        config.player_image, (player.x, player.y), level, player
+    )
 
 
 def display_monsters(screen, config, level, player):
     for monster in level.monsters:
-        image = level.images.get(monster.image)
-        screen.draw_tile_relative(image, player, (monster.x, monster.y))
+        screen.draw_tile_relative(monster.image, (monster.x, monster.y), level, player)
 
 
 def display_items(screen, config, level, player):
     for item_coords in level.item_coordinates:
         item = next((x for x in level.items if x.id == item_coords['id']), None)
-        image = level.images.get(item.image)
-        screen.draw_tile_relative(image, player, item_coords['coordinates'])
+        screen.draw_tile_relative(
+            item.image, item_coords['coordinates'], level, player
+        )
 
 
 def display_player_items(screen, config, level, player):
     screen.draw_text(
-        config.fonts.get(config.score_font), 'Inventory:', (0, 420), 'black'
+        config.score_font, 'Inventory:', (0, 420), 'black'
     )
     for index, item in enumerate(player.items):
-        image = level.images.get(item.image, None)
         # Position each item from left to right with respect to ordering and
         # on the bottom tile, adjusted by 8 for some padding
         x = index * screen.tile_width
         y = screen.height - screen.tile_height - 8
-        screen.draw(image, (x, y))
+        screen.draw(item.image, (x, y))
 
 
 def draw_popup(screen, config, level, player, strings):
@@ -54,7 +53,7 @@ def draw_popup(screen, config, level, player, strings):
     box_x, box_y = config.popup_box['x'], config.popup_box['y']
     for index, string in enumerate(strings):
         screen.draw_text(
-            config.fonts.get(config.score_font), string,
+            config.score_font, string,
             (x_margin, y_margin + index * 20), surface=message_surface
         )
         screen.draw(message_surface, (box_x, box_y))
@@ -70,7 +69,8 @@ def draw_splash(screen, config, image):
     screen.draw(message_surface, origin)
 
 
-def render(game_state, screen, config, level, player):
+def render(game_data, questions, level, player):
+    game_state, screen, config = game_data.state, game_data.screen, game_data.config
     with screen.display_cycle():
         display_map(screen, config, level, player)
         display_items(screen, config, level, player)
@@ -80,14 +80,12 @@ def render(game_state, screen, config, level, player):
         if game_state.is_state('question'):
             draw_popup(
                 screen, config, level, player,
-                config.questions.get_question_display()
+                questions.get_question_display()
             )
         if game_state.is_state('splash'):
-            image = config.images.get(config.splash_image, None)
-            draw_splash(screen, config, image)
+            draw_splash(screen, config, config.splash_image)
         if game_state.is_state('endscreen'):
-            image = config.images.get(config.endscreen_image, None)
-            draw_splash(screen, config, image)
+            draw_splash(screen, config, config.endscreen_image)
         if game_state.is_state('item'):
             draw_popup(screen, config, level, player, player.current_item.message)
         if game_state.is_state('info'):