+++ src/yape/dispatch.py
@@ -1,5 +1,4 @@
 import pygame
-from fsm import RequiresStateMachine
 
 
 class Dispatch(object):
@@ -9,37 +8,36 @@ class Dispatch(object):
     """
 
     listeners = {}
-    state_machine = None
 
     def register(self, state, listener, event_type=None):
         """
-        Given a `state` name and a `listener` function, registers the listener
-        for calling when the game is in that state.
+        Given a `state` name, a `listener` function, and an `event_type`,
+        registers the listener for calling when the game is in the given state
+        and the event type occurs.
         """
         event_type = event_type or pygame.constants.KEYDOWN
         existing = self.listeners.get((state, event_type), [])
         existing.append(listener)
         self.listeners[(state, event_type)] = existing
 
-    def attach_state_machine(self, state_machine):
-        self.state_machine = state_machine
-
-    def handle_events(self, *args, **kwargs):
-        if not self.state_machine:
-            raise RequiresStateMachine(
-                'Dispatcher has no state machine attached'
-            )
+    def handle_events(self, game_data, *args, **kwargs):
+        """
+        Given the `game_data` and any args/kwargs, dispatch pygame events to
+        the registered listeners for the current state based on the event_type
+        """
         for event in pygame.event.get():
-            self.dispatch(event, *args, **kwargs)
+            self.dispatch(event, game_data, *args, **kwargs)
 
-    def dispatch(self, event, *args, **kwargs):
-        if not self.state_machine:
-            raise RequiresStateMachine(
-                'Dispatcher has no state machine attached'
-            )
-        state = self.state_machine.state
-        for listener in self.listeners.get((state, event.type), []):
-            listener(event, self.state_machine, *args, **kwargs)
+    def dispatch(self, event, game_data, *args, **kwargs):
+        """
+        Given the global state_machine, a pygame event, and any args/kwargs,
+        call any registers that are registered for the state_machine's current
+        state paired with the event type. Any args/kwargs are passed to the
+        listeners
+        """
+        listeners = self.listeners.get((game_data.state.state, event.type), [])
+        for listener in listeners:
+            listener(event, game_data, *args, **kwargs)
 
     def register_listener(self, states, event_type=None):
         """
@@ -50,15 +48,12 @@ class Dispatch(object):
         """
 
         def decorator(f):
-
             if not getattr(f, 'registered', False):
                 for state in states:
                     self.register(state, f, event_type)
                 f.registered = True
-
-            def _(*args, **kwargs):
-                f(*args, **kwargs)
-            return _
+            return f
         return decorator
 
 dispatcher = Dispatch()
+